+++ src/fastman/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.4"
+__version__ = "0.3.5"
 __codename__ = "Cheetah"
 
 from .commands.base import Command, register